Add tests for Venn views.
These tests currently fail for the venn subset views and for the
venn_pfam view because they are broken and do not show the page title.
Moreover the tests fail for the venn_amr view which does not exist yet.
Also note that the table is broken in the ko_venn_subset view and does
not show any data.