Improve OrthogroupViewMixin to fix GWAS view for orthogroups.
For orthogroups, the get_hit_descriptions method did not return
a dataframe with one entry per hit, contrary to the same function
for all other mixins. We therefore modify it to include processing
of the data returned by the SQL query so that it returns a
DataFrame similar to get_hit_descriptions from other mixins.