Fix Scoary-2 integration.
Scoary-2 uses a database to store scores for topologies and it
caches the connection and cursor to that database when scoary gets
loaded. This leads to issues for the integration as the SQL object
might be used in a different thread than it was created in.
We fix it here by resetting the connection before calling scoary.