Fix binding of folders when running with singularity.
When running with singularity the current working directory and
home folders get mounted by default. It seems that this can collide
with our own mounts, which get mounted into webapp/assets of the
zDB directory, which usually also is the CWD. I did not quite
understand why this works in certain cases but not in others, but
it seems this led to zDB not running with singularity on our server.
Adding the -c option to avoid sharing the filesystem resolves this
issue.