Add missing assets when running webapp with nginx and gunicorn.
The dal and dal_select2 apps come with static files which need to
be served. This is done automatically when running the dev_server
but not when using nginx and gunicorn. In that case we need to
use the collectstatic command to gather all the necessary static
files into one folder and then serve that folder.
We also improve cleaning-up of the files when the webapp gets
shut-down, which is mainly an issue with docker as folders and files
get created by root, leading to issues when running with conda or
singularity after running with docker.