Optimize order of regions in plot_region.
We automatically determine the best order to display the regions
in the plot_region view, maximizing the number and order of
common orthogroups between consecutive regions. Note that we use
a local optimum, taking a path of maximum score between neighbouring
regions, not over the full path (although we take the best path
among such paths obtained for all starting regions).