Use up to date KEGG database as reference.
As zDB's base database is now created with "zdb setup", the KO
definitions it contains will be up to date, so that we also need
to have an up to date version of the KEGG database as reference.
This means that the base database and reference databases should
be created together to ensure compatibility between the two.