Make sure to add run to list of completed run upon import.
When not specifying a run name during export, so that the latest
run was used, only the latest run was updated upon import, without
referencing the actual run in the list of completed run. This meant
that if another run was completed in the same directory the imported
one would not be listed anymore.