Remove GI cluster table and add GI description table instead.
The genomic_island_clusters table was only adding complexity with
no gains, as each GI is in a single cluster. Instead we add a
genomic_island_descriptions table, which allows to have a similar
table layout for genomic islands as for other annotation types. This
will make comparison views much easier to create. Also it allows us
to store some information about the clusters in a table and not
compute them every time we need them.