Update annotation pipeline tests.
With the new orthofinder version, there are no core orthogroups left
in our test dataset, which leads to an error. It seems that there was
just one core orthogroup before and it gets merged with another OG
with orthofinder 3, leading to the issue described above. We therefore
add the his operon to one of the test genomes, so that all three
have that operon.