Improve style of circos map.
We draw both strands on the inside of the backbone, so that we can
increase the space between tracks without it increasing space between
each strand and the backbon. This allows us to remove the separation
lines between tracks and slots. We also set the colors for most tracks,
especially so that the colors used do not change when switching coloring
scheme for the reference genome. We also make sure that the legend order
matches the track order of the plot.