Make sure the refseq DB gets copied to reference database directory.
The "move" mode does not work for things passed as inputs to a
process, as those will be symlings and move does not follow them,
simply moving the symlink instead. Moreover move cannot be used
on files that are still needed by other processes, so we have
no other choice than copying the database to the reference
database folder.